Refactor: 파일 디렉토리 재배치로 인한 필요 의존성 추가
