+++ src/test/java/com/jerry/ticketing/application/seat/unit/SeatBlockingSchedulerTest.java
@@ -1,13 +1,13 @@
-package com.jerry.ticketing.application.seat;
+package com.jerry.ticketing.application.seat.unit;
 
+import com.jerry.ticketing.application.seat.SeatBlockingScheduler;
+import com.jerry.ticketing.application.seat.SeatBlockingService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
